+++ src/main.cpp
@@ -1,10 +1,12 @@
 #include "util/Badges.hpp"
+#include "util/BadgeInfo.hpp"
 
 #include <Geode/Geode.hpp>
 
 #include <Geode/utils/web.hpp>
 
 #include <Geode/modify/ProfilePage.hpp>
+#include <Geode/modify/CommentCell.hpp>
 
 using namespace geode::prelude;
 
@@ -12,16 +14,14 @@ class $modify(Profile, ProfilePage)
 {
 	struct Fields
 	{
-		EventListener<web::WebTask> ogdBadgeRequest;
 		std::string badgeID;
+		EventListener<web::WebTask> ogdBadgeRequest;
 		int accID;
 	};
 
 	void loadPageFromUserInfo(GJUserScore * user)
 	{
 		ProfilePage::loadPageFromUserInfo(user);
-
-		auto layer = m_mainLayer;
 		m_fields->accID = user->m_accountID;
 
 		Profile::onCheckForBadge();
@@ -36,7 +36,8 @@ class $modify(Profile, ProfilePage)
 		{
 			log::debug("Fetched badge {} from cache", badgeCache);
 
-			Profile::setNewBadge(badgeCache);
+			CCMenu *username_menu = typeinfo_cast<CCMenu *>(m_mainLayer->getChildByIDRecursive("username-menu"));
+			Profile::setNewBadge(badgeCache, username_menu);
 		};
 
 		m_fields->ogdBadgeRequest.bind([this](web::WebTask::Event *e)
@@ -50,8 +51,10 @@ class $modify(Profile, ProfilePage)
                 } else {
                     geode::getMod()->setSavedValue(fmt::format("cache-badge-u{}", m_fields->accID), ogdWebResultUnwrapped);
                     log::debug("Fetched badge {} remotely", ogdWebResultUnwrapped);
+					
+			CCMenu *username_menu = typeinfo_cast<CCMenu *>(m_mainLayer->getChildByIDRecursive("username-menu"));
 
-                    Profile::setNewBadge(ogdWebResultUnwrapped);
+                    Profile::setNewBadge(ogdWebResultUnwrapped, username_menu);
                 };
 			}
 			else if (web::WebProgress *p = e->getProgress())
@@ -67,11 +70,8 @@ class $modify(Profile, ProfilePage)
 		m_fields->ogdBadgeRequest.setFilter(ogdReq.get(fmt::format("https://raw.githubusercontent.com/CubicCommunity/InstallOGDPS/main/data/publicBadges/{}.txt", (int)m_fields->accID)));
 	};
 
-	void setNewBadge(std::string id)
+	void setNewBadge(std::string id, CCMenu * username_menu)
 	{
-		auto layer = m_mainLayer;
-		CCMenu *username_menu = typeinfo_cast<CCMenu *>(layer->getChildByIDRecursive("username-menu"));
-
 		if (id == Badges::getBadgeStringID[Badges::BadgeID::Director])
 		{
 			if (auto alreadyBadge = username_menu->getChildByID(id))
@@ -79,15 +79,13 @@ class $modify(Profile, ProfilePage)
 				alreadyBadge->removeMeAndCleanup();
 			};
 
-			m_fields->badgeID = id;
-
 			CCSprite *badgeSprite = CCSprite::create("crew-member.png"_spr);
 			badgeSprite->setScale(0.925f);
 
 			CCMenuItemSpriteExtra *badge = CCMenuItemSpriteExtra::create(
 				badgeSprite,
 				this,
-				menu_selector(Profile::onInfoBadge));
+				menu_selector(BadgeInfo::onInfoBadge));
 			badge->setID(id);
 			badge->setZOrder(1);
 
@@ -101,15 +99,13 @@ class $modify(Profile, ProfilePage)
 				alreadyBadge->removeMeAndCleanup();
 			};
 
-			m_fields->badgeID = id;
-
 			CCSprite *badgeSprite = CCSprite::create("team-member.png"_spr);
 			badgeSprite->setScale(0.925f);
 
 			CCMenuItemSpriteExtra *badge = CCMenuItemSpriteExtra::create(
 				badgeSprite,
 				this,
-				menu_selector(Profile::onInfoBadge));
+				menu_selector(BadgeInfo::onInfoBadge));
 			badge->setID(id);
 			badge->setZOrder(1);
 
@@ -123,15 +119,13 @@ class $modify(Profile, ProfilePage)
 				alreadyBadge->removeMeAndCleanup();
 			};
 
-			m_fields->badgeID = id;
-
 			CCSprite *badgeSprite = CCSprite::create("team-member.png"_spr);
 			badgeSprite->setScale(0.925f);
 
 			CCMenuItemSpriteExtra *badge = CCMenuItemSpriteExtra::create(
 				badgeSprite,
 				this,
-				menu_selector(Profile::onInfoBadge));
+				menu_selector(BadgeInfo::onInfoBadge));
 			badge->setID(id);
 			badge->setZOrder(1);
 
@@ -145,15 +139,13 @@ class $modify(Profile, ProfilePage)
 				alreadyBadge->removeMeAndCleanup();
 			};
 
-			m_fields->badgeID = id;
-
 			CCSprite *badgeSprite = CCSprite::create("crew-member.png"_spr);
 			badgeSprite->setScale(0.925f);
 
 			CCMenuItemSpriteExtra *badge = CCMenuItemSpriteExtra::create(
 				badgeSprite,
 				this,
-				menu_selector(Profile::onInfoBadge));
+				menu_selector(BadgeInfo::onInfoBadge));
 			badge->setID(id);
 			badge->setZOrder(1);
 
@@ -167,15 +159,13 @@ class $modify(Profile, ProfilePage)
 				alreadyBadge->removeMeAndCleanup();
 			};
 
-			m_fields->badgeID = id;
-
 			CCSprite *badgeSprite = CCSprite::create("crew-member.png"_spr);
 			badgeSprite->setScale(0.925f);
 
 			CCMenuItemSpriteExtra *badge = CCMenuItemSpriteExtra::create(
 				badgeSprite,
 				this,
-				menu_selector(Profile::onInfoBadge));
+				menu_selector(BadgeInfo::onInfoBadge));
 			badge->setID(id);
 			badge->setZOrder(1);
 
@@ -183,78 +173,170 @@ class $modify(Profile, ProfilePage)
 			username_menu->updateLayout();
 		};
 	};
+};
 
-	void onInfoBadge(CCObject * sender)
+class $modify(Comment, CommentCell)
+{
+	struct Fields
 	{
-		if (m_fields->badgeID == Badges::getBadgeStringID[Badges::BadgeID::Director])
+		std::string badgeID;
+		EventListener<web::WebTask> ogdBadgeRequest;
+		int accID;
+	};
+
+	void loadFromComment(GJComment * comment)
+	{
+		CommentCell::loadFromComment(comment);
+		m_fields->accID = comment->m_accountID;
+
+		Comment::onCheckForBadge();
+		log::debug("Viewing profile of ID {}", m_fields->accID);
+	};
+
+	void onCheckForBadge()
+	{
+		auto badgeCache = geode::getMod()->getSavedValue<std::string>(fmt::format("cache-badge-u{}", m_fields->accID)).c_str();
+
+		if (badgeCache)
 		{
-			geode::createQuickPopup(
-				"OBSIDIAN Director",
-				"This user is the <cj>director</c> of <cp>OBSIDIAN</c>. They manage and supervise everything.",
-				"OK", "Learn More",
-				[](auto, bool btn2)
-				{
-					if (btn2)
-					{
-						web::openLinkInBrowser("https://www.obsidianmg.cc/#crew");
-					};
-				});
+			log::debug("Fetched badge {} from cache", badgeCache);
+
+			CCMenu *username_menu = typeinfo_cast<CCMenu *>(m_mainLayer->getChildByIDRecursive("username-menu"));
+			Comment::setNewBadge(badgeCache, username_menu);
+		};
+
+		m_fields->ogdBadgeRequest.bind([this](web::WebTask::Event *e)
+									   {
+				if (web::WebResponse *ogdReqRes = e->getValue())
+			{
+
+				std::string ogdWebResultUnwrapped = ogdReqRes->string().unwrapOr("Uh oh!");
+
+                if (ogdWebResultUnwrapped.c_str() == geode::getMod()->getSavedValue<std::string>(fmt::format("cache-badge-u{}", m_fields->accID)).c_str()) {
+                    log::debug("Badge for user of ID {} up-to-date", m_fields->accID);
+                } else {
+                    geode::getMod()->setSavedValue(fmt::format("cache-badge-u{}", m_fields->accID), ogdWebResultUnwrapped);
+                    log::debug("Fetched badge {} remotely", ogdWebResultUnwrapped);
+					
+			CCMenu *username_menu = typeinfo_cast<CCMenu *>(m_mainLayer->getChildByIDRecursive("username-menu"));
+
+                    Comment::setNewBadge(ogdWebResultUnwrapped, username_menu);
+                };
+			}
+			else if (web::WebProgress *p = e->getProgress())
+			{
+				log::debug("badge id progress: {}", p->downloadProgress().value_or(0.f));
+			}
+			else if (e->isCancelled())
+			{
+				log::debug("The request was cancelled... So sad :(");
+			}; });
+
+		auto ogdReq = web::WebRequest();
+		m_fields->ogdBadgeRequest.setFilter(ogdReq.get(fmt::format("https://raw.githubusercontent.com/CubicCommunity/InstallOGDPS/main/data/publicBadges/{}.txt", (int)m_fields->accID)));
+	};
+
+	void setNewBadge(std::string id, CCMenu * username_menu)
+	{
+		if (id == Badges::getBadgeStringID[Badges::BadgeID::Director])
+		{
+			if (auto alreadyBadge = username_menu->getChildByID(id))
+			{
+				alreadyBadge->removeMeAndCleanup();
+			};
+
+			CCSprite *badgeSprite = CCSprite::create("crew-member.png"_spr);
+			badgeSprite->setScale(0.925f);
+
+			CCMenuItemSpriteExtra *badge = CCMenuItemSpriteExtra::create(
+				badgeSprite,
+				this,
+				menu_selector(BadgeInfo::onInfoBadge));
+			badge->setID(id);
+			badge->setZOrder(1);
+
+			username_menu->addChild(badge);
+			username_menu->updateLayout();
 		}
-		else if (m_fields->badgeID == Badges::getBadgeStringID[Badges::BadgeID::TeamManager])
+		else if (id == Badges::getBadgeStringID[Badges::BadgeID::TeamManager])
 		{
-			geode::createQuickPopup(
-				"ObsidianGD Manager",
-				"This user is a <cy>manager</c> of <cp>ObsidianGD</c>. They manage group projects and collaborations.",
-				"OK", "Learn More",
-				[](auto, bool btn2)
-				{
-					if (btn2)
-					{
-						web::openLinkInBrowser("https://www.obsidianmg.cc/gd-team/#team");
-					};
-				});
+			if (auto alreadyBadge = username_menu->getChildByID(id))
+			{
+				alreadyBadge->removeMeAndCleanup();
+			};
+
+			CCSprite *badgeSprite = CCSprite::create("team-member.png"_spr);
+			badgeSprite->setScale(0.925f);
+
+			CCMenuItemSpriteExtra *badge = CCMenuItemSpriteExtra::create(
+				badgeSprite,
+				this,
+				menu_selector(BadgeInfo::onInfoBadge));
+			badge->setID(id);
+			badge->setZOrder(1);
+
+			username_menu->addChild(badge);
+			username_menu->updateLayout();
 		}
-		else if (m_fields->badgeID == Badges::getBadgeStringID[Badges::BadgeID::TeamMember])
+		else if (id == Badges::getBadgeStringID[Badges::BadgeID::TeamMember])
 		{
-			geode::createQuickPopup(
-				"ObsidianGD Team Member",
-				"This user is a <cg>member</c> of <cp>ObsidianGD</c>. They partake in group projects and collaborations.",
-				"OK", "Learn More",
-				[](auto, bool btn2)
-				{
-					if (btn2)
-					{
-						web::openLinkInBrowser("https://www.obsidianmg.cc/gd-team/#team");
-					};
-				});
+			if (auto alreadyBadge = username_menu->getChildByID(id))
+			{
+				alreadyBadge->removeMeAndCleanup();
+			};
+
+			CCSprite *badgeSprite = CCSprite::create("team-member.png"_spr);
+			badgeSprite->setScale(0.925f);
+
+			CCMenuItemSpriteExtra *badge = CCMenuItemSpriteExtra::create(
+				badgeSprite,
+				this,
+				menu_selector(BadgeInfo::onInfoBadge));
+			badge->setID(id);
+			badge->setZOrder(1);
+
+			username_menu->addChild(badge);
+			username_menu->updateLayout();
 		}
-		else if (m_fields->badgeID == Badges::getBadgeStringID[Badges::BadgeID::CrewManager])
+		else if (id == Badges::getBadgeStringID[Badges::BadgeID::CrewManager])
 		{
-			geode::createQuickPopup(
-				"OBSIDIAN Manager",
-				"This user is a <cy>manager</c> of <cp>OBSIDIAN</c>. They manage group projects.",
-				"OK", "Learn More",
-				[](auto, bool btn2)
-				{
-					if (btn2)
-					{
-						web::openLinkInBrowser("https://www.obsidianmg.cc/#crew");
-					};
-				});
+			if (auto alreadyBadge = username_menu->getChildByID(id))
+			{
+				alreadyBadge->removeMeAndCleanup();
+			};
+
+			CCSprite *badgeSprite = CCSprite::create("crew-member.png"_spr);
+			badgeSprite->setScale(0.925f);
+
+			CCMenuItemSpriteExtra *badge = CCMenuItemSpriteExtra::create(
+				badgeSprite,
+				this,
+				menu_selector(BadgeInfo::onInfoBadge));
+			badge->setID(id);
+			badge->setZOrder(1);
+
+			username_menu->addChild(badge);
+			username_menu->updateLayout();
 		}
-		else if (m_fields->badgeID == Badges::getBadgeStringID[Badges::BadgeID::CrewMember])
+		else if (id == Badges::getBadgeStringID[Badges::BadgeID::CrewMember])
 		{
-			geode::createQuickPopup(
-				"OBSIDIAN Crew Member",
-				"This user is a <cg>member</c> of <cp>OBSIDIAN</c>. They partake in group content creation.",
-				"OK", "Learn More",
-				[](auto, bool btn2)
-				{
-					if (btn2)
-					{
-						web::openLinkInBrowser("https://www.obsidianmg.cc/#crew");
-					};
-				});
+			if (auto alreadyBadge = username_menu->getChildByID(id))
+			{
+				alreadyBadge->removeMeAndCleanup();
+			};
+
+			CCSprite *badgeSprite = CCSprite::create("crew-member.png"_spr);
+			badgeSprite->setScale(0.925f);
+
+			CCMenuItemSpriteExtra *badge = CCMenuItemSpriteExtra::create(
+				badgeSprite,
+				this,
+				menu_selector(BadgeInfo::onInfoBadge));
+			badge->setID(id);
+			badge->setZOrder(1);
+
+			username_menu->addChild(badge);
+			username_menu->updateLayout();
 		};
 	};
 };