+++ src/util/BadgeInfo.hpp
@@ -1,0 +1,16 @@
+#pragma once
+
+#include <map>
+#include <string>
+
+#include <Geode/Geode.hpp>
+
+#include <Geode/utils/web.hpp>
+
+using namespace geode::prelude;
+
+class BadgeInfo
+{
+public:
+    void onInfoBadge(CCObject *sender);
+};