+++ src/util/sources/Badges.cpp
@@ -1,5 +1,11 @@
 #include "../Badges.hpp"
 
+#include <Geode/Geode.hpp>
+
+#include <Geode/utils/web.hpp>
+
+using namespace geode::prelude;
+
 std::map<Badges::BadgeID, std::string> Badges::getBadgeStringID{
     {Badges::BadgeID::Director, "director"},
     {Badges::BadgeID::CrewManager, "crew-manager"},