+++ changelog.md
@@ -1,3 +1,8 @@
+## 1.1.0
+#### Comment badges
+
+- ![Add Icon](frame:GJ_completesIcon_001.png&scale:0.5) | **Added** comment cell compatibility
+
 # 1.0.0
 ### Initial release